wip basic SExpr tokenizer
