+++ src/main/scala/Transformer.scala
@@ -1,0 +1,15 @@
+import Expression.*
+import Literal.*
+import cats.parse.{Parser0 as P0, Parser as P, Numbers}
+import cats.data.NonEmptyList
+
+object Transformer {
+  // def toAst(expr: SExp): Expresssion = {
+  //   expr match {
+  //     case SSId(it)     => ???
+  //     case SSString(it) => ???
+  //     case SSExprs(xs)  => ???
+  //   }
+  //   ???
+  // }
+}
